git-svn-id: http://svn.resourcespace.com/svn/rs/branches/20181018_acota_q10076@12535 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/preview_preprocessing.php
@@ -252,7 +252,11 @@ if (($extension=="cr2" || $extension=="nef" || $extension=="dng" || $extension==
             // check for nef -otherimage failure
             if ($extension=="nef"&&!filesize_unlimited($target)>0)
                 {
-                unlink($target);    
+                if(file_exists($target))
+                    {
+                    unlink($target);
+                    }
+
                 $bin_tag=" -previewimage ";
                 //2nd attempt
                 $cmd=$exiftool_fullpath.' -b '.$bin_tag.' '.escapeshellarg($file).' -w %d%f.jpg';
@@ -318,7 +322,7 @@ if ( (($extension=="pages") || ($extension=="numbers") || (!isset($unoconv_path)
    ----------------------------------------
 */
 global $unoconv_extensions;
-if (in_array($extension,$unoconv_extensions) && isset($unoconv_path) && !isset($newfile))
+if (in_array($extension,$unoconv_extensions) && $extension!='pdf' && isset($unoconv_path) && !isset($newfile))
     {
     global $config_windows;
     $unocommand=$unoconv_path . "/unoconv";
@@ -332,7 +336,7 @@ if (in_array($extension,$unoconv_extensions) && isset($unoconv_path) && !isset($
             exit("Unoconv's OpenOffice Python executable not found at '$unoconv_python_path'");
             }
        }
-    $cmd=($config_windows ? $cmd_uno_python_path . ' ' : '') . $unocommand . " --format=pdf " . escapeshellarg($file);
+    $cmd=($config_windows ? escapeshellarg($cmd_uno_python_path) . ' ' : '') . escapeshellarg($unocommand) . " --format=pdf " . escapeshellarg($file);
     $output=run_command($cmd);
 
     $path_parts=pathinfo($file);
@@ -983,7 +987,24 @@ if ((!isset($newfile)) && (!in_array($extension, $ffmpeg_audio_extensions))&& (!
 # If a file has been created, generate previews just as if a JPG was uploaded.
 if (isset($newfile))
     {
-    create_previews($ref,false,"jpg",$previewonly,false,$alternative);  
-    }
+    if($GLOBALS['non_image_types_generate_preview_only'] && in_array($extension, $GLOBALS['non_image_types']))
+        {
+        $file_used_for_previewonly = get_resource_path($ref, true, "tmp", false, "jpg");
 
-?>
+        if(copy($newfile, $file_used_for_previewonly))
+            {
+            $previewonly = true;
+            debug("preview_preprocessing: changing previewonly = true for non-image file");
+            }
+        }
+
+    create_previews($ref,false,"jpg",$previewonly,false,$alternative);
+
+    if(
+        $GLOBALS['non_image_types_generate_preview_only']
+        && in_array($extension, $GLOBALS['non_image_types'])
+        && file_exists($file_used_for_previewonly))
+        {
+        unlink($file_used_for_previewonly);
+        }
+    }
